Simplify the seeds according to our needs
We don't need all the complexity of upstream Ubuntu/Kubuntu.

We only ship one ISO and we don't really have a "supported" set.

Simplify our seeds to reflect this.

Also, switch back to depending on ubuntu's "platform.precise" for
now. We probably don't want to do changes so deeply for a desktop
system after all. If things change in the future we will reevaluate
this decision.